moved recommeded farming computeds out of store (for now)
